Split gherkin writer helpers into a series of unique modules
